+++ backend/server/server/app/entradas/models.py
@@ -8,7 +8,7 @@ class Entrada(models.Model):
     asiento = models.IntegerField('asiento')
     graderia = models.CharField('graderia',max_length=100)
     fila = models.IntegerField('fila')
-    precio = models.DecimalField('precio', max_digits=100, decimal_places=100)
+    precio = models.IntegerField('precio')
     disponible = models.BooleanField('disponible')
 
     def __str__(self):
